+++ iris/service/content/city/__init__.py
@@ -1,8 +1,11 @@
 from . import service  # noqa
 from .document import City  # noqa
+from . import document
 
 
 def includeme(config):
+    settings = config.get_settings()
+    document.DEFAULT_TRESHOLD = int(settings.get('iris.city.treshold', 0))
     from iris.service.rest import auth
     config.add_route('city_public_api',
                      'cities',